# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Charity API",
        "description": "توثيق RESTful API لنظام الجمعية الخيرية",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local API Server"
        },
        {
            "url": "https://Al-Talahom-Charity.com/api",
            "description": "Production API Server"
        }
    ],
    "paths": {
...
        "/api/subscribers": {
            "get": {
                "tags": [
                    "Subscribers"
                ],
                "summary": "جلب كل المشتركين",
                "operationId": "2297b9a29b7b7cb84d47f3bcb08db0a2",
                "responses": {
                    "200": {
                        "description": "نجاح العملية",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SubscriberResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Subscribers"
                ],
                "summary": "إضافة مشترك جديد",
                "operationId": "fcf0d1c79723b3169188dd396283dd31",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "subscription_status"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "subscription_status": {
                                        "type": "string",
                                        "example": "active"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "تم الإنشاء",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SubscriberResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/subscribers/{id}": {
            "get": {
                "tags": [
                    "Subscribers"
                ],
                "summary": "عرض مشترك واحد",
                "operationId": "7bbfc8eff81c608240401f8355132ddf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "نجاح العملية",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SubscriberResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "غير موجود"
                    }
                }
            },
            "put": {
                "tags": [
                    "Subscribers"
                ],
                "summary": "تحديث مشترك",
                "operationId": "500e71e5b53c74954952eb37b917c12c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "subscription_status": {
                                        "type": "string",
                                        "example": "inactive"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "تم التحديث",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SubscriberResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "غير موجود"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Subscribers"
                ],
                "summary": "حذف مشترك",
                "operationId": "b7f24335446073af8a627c5cfe6bdd63",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "تم الحذف"
                    },
                    "404": {
                        "description": "غير موجود"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Role": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "admin"
                    },
                    "description": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "User": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Ahmed"
                    },
                    "email": {
                        "type": "string",
                        "example": "user@example.com"
                    },
                    "phone": {
                        "type": "integer",
                        "example": 77234566
                    },
                    "roles": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Role"
                        }
                    }
                },
                "type": "object"
            },
...
            "Subscriber": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    },
                    "subscription_status": {
                        "type": "string",
                        "example": "active"
                    }
                },
                "type": "object"
            },
            "SubscriberResponse": {
                "properties": {
                    "status": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "Subscribers retrieved successfully"
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Subscriber"
                        }
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}